Move e2e tests to their respective templates
In the future users will have an option to have
these tests included in their starter templates
