+++ examples/content-react/content-react.spec.ts
@@ -1,6 +1,6 @@
 import path from 'path'
 import {execSync} from 'child_process'
-import {extensionFixtures} from './extension-fixtures'
+import {extensionFixtures} from '../extension-fixtures'
 
 const exampleDir = 'examples/content-react'
 const pathToExtension = path.join(__dirname, `../${exampleDir}/dist/chrome`)
