+++ examples/content-sass/content-sass-module.spec.ts
@@ -1,6 +1,6 @@
 import path from 'path'
 import {execSync} from 'child_process'
-import {extensionFixtures} from './extension-fixtures'
+import {extensionFixtures} from '../extension-fixtures'
 
 const exampleDir = 'examples/content-sass-module'
 const pathToExtension = path.join(__dirname, `../${exampleDir}/dist/chrome`)
