+++ examples/content-typescript/content-typescript.spec.ts
@@ -1,6 +1,6 @@
 import path from 'path'
 import {execSync} from 'child_process'
-import {extensionFixtures} from './extension-fixtures'
+import {extensionFixtures} from '../extension-fixtures'
 
 const exampleDir = 'examples/content-typescript'
 const pathToExtension = path.join(__dirname, `../${exampleDir}/dist/chrome`)
