+++ examples/extension-fixtures.ts
