+++ examples/new-typescript/new-typescript.spec.ts
@@ -1,8 +1,8 @@
 import path from 'path'
 import {execSync} from 'child_process'
-import {extensionFixtures} from './extension-fixtures'
+import {extensionFixtures} from '../extension-fixtures'
 
-const exampleDir = 'examples/new-tailwind'
+const exampleDir = 'examples/new-typescript'
 const pathToExtension = path.join(__dirname, `../${exampleDir}/dist/chrome`)
 const test = extensionFixtures(pathToExtension, true)
 
@@ -16,20 +16,18 @@ test('should exist an element with the welcome message text', async ({
   page
 }) => {
   await page.goto('chrome://newtab/')
-  const h2 = page.locator('h2')
-  await test
-    .expect(h2)
-    .toHaveText('This is a new tab page running React and Tailwind.css.')
+  const h1 = page.locator('h1')
+  await test.expect(h1).toHaveText('Welcome to your TypeScript Extension.')
 })
 
 test('should exist a default color value', async ({page}) => {
   await page.goto('chrome://newtab/')
-  const h2 = page.locator('h2')
+  const h1 = page.locator('h1')
   const color = await page.evaluate(
     (locator) => {
       return window.getComputedStyle(locator!).getPropertyValue('color')
     },
-    await h2.elementHandle()
+    await h1.elementHandle()
   )
-  await test.expect(color).toEqual('rgb(255, 255, 255)')
+  await test.expect(color).toEqual('rgb(74, 74, 74)')
 })
