+++ examples/new-vue/new-vue.spec.ts
@@ -1,8 +1,8 @@
 import path from 'path'
 import {execSync} from 'child_process'
-import {extensionFixtures} from './extension-fixtures'
+import {extensionFixtures} from '../extension-fixtures'
 
-const exampleDir = 'examples/new-preact'
+const exampleDir = 'examples/new-vue'
 const pathToExtension = path.join(__dirname, `../${exampleDir}/dist/chrome`)
 const test = extensionFixtures(pathToExtension, true)
 
@@ -17,7 +17,7 @@ test('should exist an element with the welcome message text', async ({
 }) => {
   await page.goto('chrome://newtab/')
   const h1 = page.locator('h1')
-  await test.expect(h1).toHaveText('Welcome to your Preact Extension.')
+  await test.expect(h1).toHaveText('Welcome to your Vue Extension.')
 })
 
 test('should exist a default color value', async ({page}) => {
